rag/requirements-base.txt
absl-py==2.1.0
accelerate==1.2.1
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiolimiter==1.2.1
aiosignal==1.3.2
albucore==0.0.16
albumentations==1.4.11
angle-emb==0.3.10
annotated-types==0.7.0
ansi2html==1.9.2
antlr4-python3-runtime==4.9.3
anyio==4.7.0
astor==0.8.1
async-timeout==5.0.1
attrdict==2.0.1
attrs==23.2.0
babel==2.16.0
bce-python-sdk==0.9.25
beautifulsoup4==4.12.3
# bitsandbytes==0.45.0
black==24.10.0
boltons==24.1.0
boto3==1.35.92
botocore==1.35.92
braceexpand==0.1.7
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudpickle==3.1.0
cnocr==2.3.1
cnstd==1.2.5.2
#cohere==5.3.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
cryptography==44.0.0
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.11
dash==2.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.19.1
decorator==4.4.2
dependency-injector==4.41.0
#detectron2==0.6
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
doclayout_yolo==0.0.2b1
docutils==0.19
einops==0.8.0
et_xmlfile==2.0.0
eva-decord==0.6.1
eval_type_backport==0.2.2
evaluate==0.4.3
exceptiongroup==1.2.2
fairscale==0.4.13
fast-langdetect==0.2.5
fastapi==0.115.6
fastavro==1.10.0
fasttext-predict==0.9.2.4
fasttext-wheel==0.9.2
ffmpeg-python==0.2.0
filelock==3.16.1
fire==0.7.0
FlagEmbedding==1.2.9
Flask==2.2.5
flask-babel==4.0.0
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.3.1
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
gritlm==1.0.0
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
imgaug==0.4.0
iniconfig==2.0.0
interegular==0.3.3
invoke==1.7.3
iopath==0.1.9
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.0
jsonschema==4.19.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langdetect==1.0.9
lark==1.2.2
lazy_loader==0.4
Levenshtein==0.26.1
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.43.0
lmdb==1.5.1
loguru==0.7.3
lxml==5.2.1
magic-pdf==1.1.0
Markdown==3.7
markdown-it-py==2.2.0
markdownify==0.12.1
MarkupSafe==3.0.2
matplotlib==3.10.0
mdformat==0.7.21
mdformat_deflist==0.1.3
mdformat_frontmatter==2.0.1
mdformat_myst==0.1.5
mdformat_tables==1.0.0
mdit-py-plugins==0.3.5
mdurl==0.1.2
mmengine-lite==0.10.5
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
mteb==1.26.4
multidict==6.1.0
multiprocess==0.70.16
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
#nougat-ocr==0.1.17
numba==0.60.0
numpy==1.26.4
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-ml-py==12.560.30
# nvidia-nccl-cu12==2.20.5
# nvidia-nvjitlink-cu12==12.6.85
# nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==1.61.1
openai-whisper==20231117
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optimum==1.23.3
orjson==3.10.13
outlines==0.0.46
packaging==24.0
paddleocr==2.7.3
# --extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu118/
# paddlepaddle-gpu==3.0.0b1
paddlepaddle==3.0.0b1
pandas==2.2.3
pathspec==0.12.1
pdf2docx==0.5.8
pdfminer.six==20231228
peft==0.11.1
pillow==11.1.0
pix2text==1.1.1
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
polars==1.19.0
portalocker==3.1.1
premailer==3.10.0
prettytable==3.12.0
proglog==0.1.10
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.1.0
pyarrow-hotfix==0.6
pybind11==2.13.6
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMuPDF==1.24.14
PyMuPDFb==1.24.3
# pynvml==12.0.0
pyparsing==3.2.1
pypdf==5.1.0
pypdfium2==4.30.1
pyspellchecker==0.8.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytorch-lightning==2.5.0.post0
pytrec-eval-terrier==0.5.6
pytubefix==6.3.3
pytz==2024.2
PyYAML==6.0.1
rapid-table==1.0.3
RapidFuzz==3.11.0
rapidocr-onnxruntime==1.3.25
rapidocr-paddle==1.4.4
rarfile==4.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
robust-downloader==0.0.2
rpds-py==0.22.3
rst-to-myst==0.4.0
ruamel.yaml==0.18.8
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
safetensors==0.5.0
scenedetect==0.6.3
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
sconf==0.2.5
seaborn==0.13.2
sentence-transformers==3.1.1
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shapely==2.0.6
shortuuid==1.0.13
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
struct-eqtable==0.3.2
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
# tensorrt==10.7.0
# tensorrt-cu12==10.7.0
# tensorrt-cu12-bindings==10.7.0
# tensorrt-cu12-libs==10.7.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.12.12
tiktoken==0.6.0
timm==0.9.16
tokenizers==0.19.1
tomli==2.2.1
torch==2.3.0
torchvision==0.18.0
torchmetrics==1.6.1
tqdm==4.66.2
transformers==4.42.4
# triton==2.3.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.57
ultralytics-thop==2.0.13
Unidecode==1.3.8
unimernet==0.2.3
urllib3==2.3.0
uvicorn==0.34.0
visualdl==2.5.3
voyageai==0.2.2
Wand==0.6.13
wandb==0.19.1
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==2.2.3
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
